// types/marketplace-api-schema.ts
export interface AppRevisionModel {
  /**
   * Gets the unique identifier of the app revision
   */
  id?: string // uuid
  /**
   * Gets the unique identifier of the app the revision is associated with
   */
  appId?: string // uuid
  /**
   * Gets the GUID of the developer of made the app
   */
  developerId?: string // uuid
  /**
   * Gets the full name of the app revision
   */
  name?: string
  /**
   * Gets a short summary of the app revision
   */
  summary?: string
  /**
   * Gets a detailed description of the app revision
   */
  description?: string
  /**
   * Gets the name of the developer who created the app revision
   */
  developer?: string
  /**
   * Gets the email address of the developer's helpdesk responsible for providing support of the application
   */
  supportEmail?: string
  /**
   * Gets the telephone number of the developer's helpdesk responsible for providing support of the application
   */
  telephone?: string
  /**
   * Gets the home page of the developer, or the app/product specific page on the developer's website
   */
  homePage?: string
  /**
   * Gets the app revisions launch uri
   */
  launchUri?: string
  /**
   * Gets the listed status of the app revision
   */
  isListed?: boolean
  /**
   * Gets a flag determining whether or not the app will appear in the marketplace
   */
  isDirectApi?: boolean
  /**
   * Gets the app revisions category
   */
  category?: CategoryModel
  /**
   * Gets the scopes that have been associated to this revision
   */
  scopes?: ScopeModel[]
  /**
   * Gets a collection of media objects associated with the app
   */
  media?: MediaModel[]
  /**
   * Gets the links associated to this model
   */
  readonly links?: LinkModel[]
}
// ...
export interface CategoryModel {
  /**
   * Gets the unique identifier associated to the category
   */
  id?: string
  /**
   * Gets the name of the category
   */
  name?: string
  /**
   * Gets the description of the category
   */
  description?: string
  /**
   * Gets the links associated to this model
   */
  readonly links?: LinkModel[]
}
// ...
export interface LinkModel {
  /**
   * The relationship being described
   */
  rel?: string
  /**
   * The hyperlink URI
   */
  href?: string
  /**
   * The HTTP verb to be issued
   */
  action?: string
}
/**
 * Media representation
 */
export interface MediaModel {
  /**
   * Gets the unique identifier of the media item
   */
  id?: string // uuid
  /**
   * Gets the URI where this media is located
   */
  uri?: string
  /**
   * Gets the textual description of this media content
   */
  description?: string
  /**
   * Gets type of media this entity relates to
   */
  type?: string
  /**
   * Gets the order of this particular picture in the list of available media
   */
  order?: number // int32
  /**
   * Gets the links associated to this model
   */
  readonly links?: LinkModel[]
}
// ...
export interface ScopeModel {
  /**
   * Gets the name of the scope
   */
  name?: string
  /**
   * Gets the description of the scope
   */
  description?: string
}